cacheinfo: move get_cpu_cacheinfo_id() back out
This commit more or less reverts commit 709c4362725a ("cacheinfo:
Move resctrl's get_cache_id() to the cacheinfo header file").

There are no users of the static inline helper outside of resctrl/core.c
and cpu.h is a pretty heavy include, it pulls in device.h etc. This
trips up architectures like riscv which want to access cacheinfo
in low level headers like elf.h.

Link: https://lore.kernel.org/all/20211120035253.72074-1-kuba@kernel.org/
